// src/app/models/Requisiton.ts
export class Requisiton {
  // tslint:disable-next-line:variable-name
  requisition_id?: any;
  manager?: string;
  status?: string;
  type?: string;
  site: string;
  // tslint:disable-next-line:variable-name
  store_location: string;
  // tslint:disable-next-line:variable-name
  due_date: string;
}